+++ packages/interact/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/interact
 
+## 0.69.0-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.69.0-pre.0
+- @glissade/player@0.69.0-pre.0
+- @glissade/scene@0.69.0-pre.0
+
 ## 0.68.0
 
 ### Patch Changes