+++ packages/scene/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/scene
 
+## 0.69.0-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.69.0-pre.0
+
 ## 0.68.0
 
 ### Minor Changes